Added an input field for importing a tree, and added an export button that displays the current tree state as a base 64 string for exporting
The import/export functionality works perfectly with url's from GGG's site, but I haven't tested it with any of the other community tools

This commit also swaps some methods in SkillTree.jsx to use nodeId instead of a node object.

There was an in retrospect pretty obvious race condition with toggleNode that only showed up when I was importing trees.

The nodes being passed in to the updater could be out of date if a bunch of state changes were being queued up really fast, and so toggleNode could end up using the wrong adjacentChange value and breaking things.

Passing in a nodeId instead and finding that node in the state when the updater gets called fixes the race condition.

I swapped the other functions that were passing in a node to use a nodeId for consistency

Also added some more styling